feat(instructions): i64, f32 and f64 constant values

+++ src/interpreter/runtime/values/i64.js
@@ -1,6 +1,6 @@
 // @flow
 
-const type = 'i32';
+const type = 'i64';
 
 export function createValue(value: number): StackLocal {
